chore: Unsafe.java source in scala source tree #31422
